feat: Add rootFunctionIds array to store root-level function IDs
This commit modifies the `detectClassesAndFunctions` function in `detector.js` to include a new array called `rootFunctionIds`. This array is used to store the IDs of root-level functions, i.e., functions that are not nested within other functions or classes. This change enables easier access and identification of root-level functions for further processing or analysis.