feat: Add FlowUpdater component to CodeMap
This commit adds the FlowUpdater component to the CodeMap in order to update the flow chart based on the focusNodeId. The FlowUpdater component uses the useReactFlow hook to get the center position of the focused node and sets it as the center of the flow chart with a zoom level of 1.5. This improves the user experience by automatically centering and zooming in on the focused node.

Refactor: Remove 'function' from the list of function types in AST analysis
Refactor: Improve function call relationship handling and add duplicate count tracking
Got the graph working kinda
Changed test code