feat: Add react-markdown dependency and update createNode function
The commit adds the "react-markdown" dependency to the package.json file. This dependency is required for rendering Markdown content in the application.

In addition, the createNode function in the transformToReactFlowData.js file is updated to include the start and end positions of the code. This change enhances the functionality of the createNode function by providing more information about the code location.

Refactor: Remove 'function' from the list of function types in AST analysis