feat: Update analyze-directory IPC handler to include option for including anonymous functions
This commit updates the analyze-directory IPC handler in the main.js file to include an option for including anonymous functions in the analysis. The handler now accepts an additional parameter, includeAnonymousFunctions, which is passed to the analyzeDirectory function. This change enhances the functionality of the analyze-directory feature by allowing the user to choose whether or not to include anonymous functions in the analysis results.