feat: Handle anonymous function names in ElkNode component
This commit adds logic to handle anonymous function names in the ElkNode component. If the label of the node starts with 'anonymous_', it is truncated to 13 characters followed by ellipsis. This enhancement improves the readability of the node labels in the mind map.

Refactor: Remove unnecessary code in CodeMap component